chore: move erro page file
